Update: set the current item explicitly in a reducer action. Allows the ability to add higher priority cards in the background.
